MASSIVE REFACTOR, client/server skin filters, and health bars
MASSIVE overhaul over the file structure of the Lua aspect, separating code into files for easier reading and modification. This should help me (and others using this codebase) more easily update or add new code.
I also added the ability for clients to set if a skin is banned, uses shield hacks, and/or has unique death sequences and save it to a file.
Servers get the ability to server-ban skins, but these settings aren't saved quite yet, so it's best that server owners keep a config file that bans the skins upon boot.
The server list can be disabled globally via "hellfire_useSrvList", or admins can allow individual players to bypass it via "hellfire_admin bypass <name>".
There are now health bars above every player's head that will show how many hits they have left, along with the progress of regaining a hit.
Health bars will follow the client's skin setting, and the client can only see other's health bars, as they don't need to see their own as they can just look at their HUD.
Health bars can be hidden client-side with the "healthbars" variable in the set command, and can be disabled server-wide with the "hellfire_allowBars" cvar.
There has also been adjustments to the help text for some of the commands, take a look!
Also, more checks have been put in place to avoid errors, and the HUD elements of this mod won't draw when spectating.
This is a HUGE update that I was originally going to save for the release of SRB2 2.2.14, but I realized that I should release this earlier to prepare the code for the big update I have planned for 2.2.14.
A update video that goes into these changes will release with this V2 update sometime soon.